add selectors
